functions/sedo_execution-processor/requirements.txt
boto3==1.21.46
jsonschema==4.4.0
pyyaml==6.0
python-dateutil==2.8.2